Made public R.oo play nice with NAMESPACE
Apparently I need to export both the generic function and the fully
qualified name for each class, or else the R.oo syntactic sugar won't
work.  This commit doesn't represent a working state, because I need to
fix the private files, but it should be OK for the public ones.