Made script to reproduce steelStrain from raw data
I want my data generation to be more reproducible and transparent, so I
made a script that does what I want it to with the raw data for the
steelStrain dataset.  It will be tracked in github, but the datafile
will be excluded from the package build (it's 5.1 MB; that's too big).